// types/crm.ts
export type CRMData = {
    name: string | null;
    phonenum: string | null;
    phoneoperator: string | null;
    bookname?: string | null;
    bookauthor?: string | null;
    file?: string | null;
    url?: string | null;
    bounding: string | null;
    color: string | null;
    advancedservices: string | null;
    comment: string | null;
}